+++ package.json
@@ -34,6 +34,7 @@
     "css-loader": "0.28.7",
     "dotenv": "4.0.0",
     "dotenv-expand": "4.2.0",
+    "enzyme": "^3.3.0",
     "eslint": "4.10.0",
     "eslint-config-airbnb-base": "^11.3.2",
     "eslint-config-getninjas": "^4.0.0",
