+++ spec/components/Button.spec.js
@@ -1,0 +1,35 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import { shallow } from 'enzyme';
+import Button from '../../src/components/Button';
+
+describe('Button', () => {
+  it('renders custom props', () => {
+    const component = renderer.create(
+      <Button
+        isSubmit={false}
+        handleButtonClick={()=>{}}
+      >Test</Button>,
+    );
+
+    const tree = component.toJSON();
+
+    expect(tree).toMatchSnapshot();
+  });
+
+  it('renders defaultProps', () => {
+    const component = renderer.create(<Button handleButtonClick={() => { }}>Test</Button>);
+
+    const tree = component.toJSON();
+
+    expect(tree).toMatchSnapshot();
+  });
+
+  describe('when prop submit is true', () => {
+    it('renders button with type=submit', () => {
+      const component = shallow(<Button isSubmit={true} handleButtonClick={() => { }}>Test</Button>);
+
+      expect(component.type()).to.equal('submit');
+    });
+  });
+});
