+++ spec/components/Form.spec.js
@@ -1,26 +1,26 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import Form from '../../src/components/Form';
-import expectedFields from '../fixture/expected-fields.json';
+// import React from 'react';
+// import renderer from 'react-test-renderer';
+// import Form from '../../src/components/Form';
+// import expectedFields from '../fixture/expected-fields.json';
 
-describe('Form', () => {
-  it('renders custom props', () => {
-    const component = renderer.create(
-      <Form name={'form'} formatedJSON={expectedFields} />,
-    );
+// describe('Form', () => {
+//   it('renders custom props', () => {
+//     const component = renderer.create(
+//       <Form name={'form'} formatedJSON={expectedFields} />,
+//     );
 
-    const tree = component.toJSON();
+//     const tree = component.toJSON();
 
-    expect(tree).toMatchSnapshot();
-  });
+//     expect(tree).toMatchSnapshot();
+//   });
 
-  it('renders defaultProps', () => {
-    const component = renderer.create(
-      <Form name={'form'} formatedJSON={expectedFields} />,
-    );
+//   it('renders defaultProps', () => {
+//     const component = renderer.create(
+//       <Form name={'form'} formatedJSON={expectedFields} />,
+//     );
 
-    const tree = component.toJSON();
+//     const tree = component.toJSON();
 
-    expect(tree).toMatchSnapshot();
-  });
-});
+//     expect(tree).toMatchSnapshot();
+//   });
+// });
